Added database backend.

+++ SharpChat/ChatEventManager.cs
@@ -1,13 +1,12 @@
 ﻿using SharpChat.Events;
 using SharpChat.Packet;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
 namespace SharpChat {
     public class ChatEventManager : IDisposable {
-        private readonly List<IChatEvent> Events = new List<IChatEvent>();
+        private readonly List<IChatEvent> Events = null;
 
         public readonly ChatContext Context;
 
@@ -15,50 +14,70 @@ namespace SharpChat {
 
         public ChatEventManager(ChatContext context) {
             Context = context;
+
+            if (Database.HasDatabase)
+                Events = new List<IChatEvent>();
         }
 
         public void Add(IChatEvent evt) {
             if (evt == null)
                 throw new ArgumentNullException(nameof(evt));
 
-            lock(Events)
-                Events.Add(evt);
+            if(Events != null)
+                lock(Events)
+                    Events.Add(evt);
 
-            Database.LogEvent(evt);
+            if(Database.HasDatabase)
+                Database.LogEvent(evt);
         }
 
         public void Remove(IChatEvent evt) {
             if (evt == null)
                 return;
 
-            lock (Events)
-                Events.Remove(evt);
+            if (Events != null)
+                lock (Events)
+                    Events.Remove(evt);
 
-            Database.DeleteEvent(evt);
+            if (Database.HasDatabase)
+                Database.DeleteEvent(evt);
+
             Context.Send(new ChatMessageDeletePacket(evt.SequenceId));
         }
 
-        public IChatEvent Get(int seqId) {
+        public IChatEvent Get(long seqId) {
             if (seqId < 1)
                 return null;
 
-            lock (Events)
-                return Events.FirstOrDefault(e => e.SequenceId == seqId);
+            if (Database.HasDatabase)
+                return Database.GetEvent(seqId);
+
+            if (Events != null)
+                lock (Events)
+                    return Events.FirstOrDefault(e => e.SequenceId == seqId);
+
+            return null;
         }
 
         public IEnumerable<IChatEvent> GetTargetLog(IPacketTarget target, int amount = 20, int offset = 0) {
-            lock (Events) {
-                IEnumerable<IChatEvent> subset = Events.Where(e => e.Target == target || e.Target == null);
+            if (Database.HasDatabase)
+                return Database.GetEvents(target, amount, offset).Reverse();
 
-                int start = subset.Count() - offset - amount;
+            if (Events != null)
+                lock (Events) {
+                    IEnumerable<IChatEvent> subset = Events.Where(e => e.Target == target || e.Target == null);
 
-                if(start < 0) {
-                    amount += start;
-                    start = 0;
+                    int start = subset.Count() - offset - amount;
+
+                    if(start < 0) {
+                        amount += start;
+                        start = 0;
+                    }
+
+                    return subset.Skip(start).Take(amount).ToList();
                 }
 
-                return subset.Skip(start).Take(amount).ToList();
-            }
+            return Enumerable.Empty<IChatEvent>();
         }
 
         ~ChatEventManager()
